Make react-dnd dependency optional
