+++ packages/react-querybuilder/package.json
@@ -51,6 +51,8 @@
     "babel-plugin-istanbul": "^6.0.0",
     "copyfiles": "^2.4.1",
     "react": "^18.1.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.1.0",
     "replace": "^1.2.1",
     "rimraf": "^3.0.2",
@@ -59,13 +61,20 @@
     "vite-tsconfig-paths": "^3.3.17"
   },
   "dependencies": {
-    "immer": "^9.0.6",
-    "react-dnd": "^16.0.1",
-    "react-dnd-html5-backend": "^16.0.1"
+    "immer": "^9.0.6"
   },
   "peerDependencies": {
     "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
+    "react-dnd": ">=14.0.0",
+    "react-dnd-html5-backend": ">=14.0.0"
+  },
+  "peerDependenciesMeta": {
+    "react-dnd": {
+      "optional": true
+    },
+    "react-dnd-html5-backend": {
+      "optional": true
+    }
   },
   "browserslist": "> 0.25%, not dead",
   "scripts": {
