+++ packages/react-querybuilder/src/Rule.test.tsx
@@ -1,12 +1,6 @@
-import { act, cleanup, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { forwardRef } from 'react';
-import {
-  simulateDrag,
-  simulateDragDrop,
-  simulateDragHover,
-  wrapWithTestBackend,
-} from 'react-dnd-test-utils';
 import { defaultControlElements } from './controls';
 import {
   defaultControlClassnames,
@@ -15,7 +9,8 @@ import {
   TestID,
 } from './defaults';
 import { errorDeprecatedRuleProps } from './internal';
-import { Rule as RuleOriginal } from './Rule';
+import { dndFallback } from './internal/hooks';
+import { Rule } from './Rule';
 import type {
   ActionProps,
   Classnames,
@@ -33,13 +28,6 @@ import type {
 } from './types';
 
 const user = userEvent.setup();
-
-const [Rule, getDndBackendOriginal] = wrapWithTestBackend(RuleOriginal);
-// This is just a type guard against `undefined`
-const getDndBackend = () => getDndBackendOriginal()!;
-
-const getHandlerId = (el: HTMLElement, dragDrop: 'drag' | 'drop') => () =>
-  el.getAttribute(`data-${dragDrop}monitorid`);
 
 const consoleError = jest.spyOn(console, 'error').mockImplementation(() => {});
 const consoleWarn = jest.spyOn(console, 'warn').mockImplementation(() => {});
@@ -136,13 +124,14 @@ const schema: Partial<Schema> = {
   ],
   showCloneButtons: false,
   validationMap: {},
+  dnd: dndFallback,
 };
 const actions: Partial<QueryActions> = {
   onPropChange: () => {},
   onRuleRemove: () => {},
 };
 const UNUSED = 'UNUSED';
-const getProps = (
+export const getProps = (
   mergeIntoSchema: Partial<Schema> = {},
   mergeIntoActions: Partial<QueryActions> = {}
 ): RuleProps => ({
@@ -292,93 +281,6 @@ describe('validation', () => {
   });
 });
 
-describe('enableDragAndDrop', () => {
-  afterEach(() => {
-    cleanup();
-  });
-
-  it('should not have the drag class if not dragging', () => {
-    render(<Rule {...getProps()} />);
-    const rule = screen.getByTestId(TestID.rule);
-    expect(rule).not.toHaveClass(sc.dndDragging);
-  });
-
-  it('should have the drag class if dragging', () => {
-    render(<Rule {...getProps()} />);
-    const rule = screen.getByTestId(TestID.rule);
-    simulateDrag(getHandlerId(rule, 'drag'), getDndBackend());
-    expect(rule).toHaveClass(sc.dndDragging);
-    act(() => {
-      getDndBackend().simulateEndDrag();
-    });
-  });
-
-  it('should have the over class if hovered', () => {
-    render(
-      <div>
-        <Rule {...getProps()} path={[0]} />
-        <Rule {...getProps()} path={[1]} />
-      </div>
-    );
-    const rules = screen.getAllByTestId(TestID.rule);
-    simulateDragHover(
-      getHandlerId(rules[0], 'drag'),
-      getHandlerId(rules[1], 'drop'),
-      getDndBackend()
-    );
-    expect(rules[1]).toHaveClass(sc.dndOver);
-    act(() => {
-      getDndBackend().simulateEndDrag();
-    });
-  });
-
-  it('should handle a dropped rule', () => {
-    const moveRule = jest.fn();
-    render(
-      <div>
-        <Rule {...getProps({}, { moveRule })} path={[0]} />
-        <Rule {...getProps({}, { moveRule })} path={[1]} />
-      </div>
-    );
-    const rules = screen.getAllByTestId(TestID.rule);
-    simulateDragDrop(
-      getHandlerId(rules[0], 'drag'),
-      getHandlerId(rules[1], 'drop'),
-      getDndBackend()
-    );
-    expect(rules[0]).not.toHaveClass(sc.dndDragging);
-    expect(rules[1]).not.toHaveClass(sc.dndOver);
-    expect(moveRule).toHaveBeenCalledWith([0], [2]);
-  });
-
-  it('should abort move if dropped on itself', () => {
-    const moveRule = jest.fn();
-    render(<Rule {...getProps({}, { moveRule })} />);
-    const rule = screen.getByTestId(TestID.rule);
-    simulateDragDrop(getHandlerId(rule, 'drag'), getHandlerId(rule, 'drop'), getDndBackend());
-    expect(rule).not.toHaveClass(sc.dndDragging);
-    expect(rule).not.toHaveClass(sc.dndOver);
-    expect(moveRule).not.toHaveBeenCalled();
-  });
-
-  it('should not try to update query if disabled', () => {
-    const moveRule = jest.fn();
-    render(
-      <div>
-        <Rule {...getProps({}, { moveRule })} path={[0]} />
-        <Rule {...getProps({}, { moveRule })} path={[1]} disabled />
-      </div>
-    );
-    const rules = screen.getAllByTestId(TestID.rule);
-    simulateDragDrop(
-      getHandlerId(rules[0], 'drag'),
-      getHandlerId(rules[1], 'drop'),
-      getDndBackend()
-    );
-    expect(moveRule).not.toHaveBeenCalled();
-  });
-});
-
 describe('disabled', () => {
   it('should have the correct classname', () => {
     render(<Rule {...getProps()} disabled />);
@@ -441,23 +343,6 @@ describe('locked rule', () => {
     render(<Rule {...getProps({ showLockButtons: true }, { onPropChange })} disabled />);
     await user.click(screen.getByTestId(TestID.lockRule));
     expect(onPropChange).toHaveBeenCalledWith('disabled', false, [0]);
-  });
-
-  it('prevents drops', () => {
-    const moveRule = jest.fn();
-    render(
-      <div>
-        <Rule {...getProps({}, { moveRule })} path={[0]} disabled />
-        <Rule {...getProps({}, { moveRule })} path={[1]} />
-      </div>
-    );
-    const rules = screen.getAllByTestId(TestID.rule);
-    simulateDragDrop(
-      getHandlerId(rules[1], 'drag'),
-      getHandlerId(rules[0], 'drop'),
-      getDndBackend()
-    );
-    expect(moveRule).not.toHaveBeenCalled();
   });
 });
