+++ packages/react-querybuilder/src/RuleGroup.tsx
@@ -1,11 +1,11 @@
-import { Fragment, useRef, type MouseEvent as ReactMouseEvent } from 'react';
-import { useDrag, useDrop } from 'react-dnd';
-import { defaultCombinators, DNDType, standardClassnames, TestID } from './defaults';
+import { Fragment, type MouseEvent as ReactMouseEvent } from 'react';
+import { defaultCombinators, standardClassnames, TestID } from './defaults';
 import { InlineCombinator } from './InlineCombinator';
+import { InlineCombinatorDnD } from './InlineCombinatorDnD';
 import { c, getValidationClassNames } from './internal';
 import { useDeprecatedProps } from './internal/hooks';
-import type { DraggedItem, RuleGroupProps } from './types';
-import { getParentPath, isAncestor, pathsAreEqual } from './utils';
+import type { RuleGroupProps } from './types';
+import { getParentPath, pathsAreEqual } from './utils';
 
 export const RuleGroup = ({
   id,
@@ -45,6 +45,11 @@ export const RuleGroup = ({
     showLockButtons,
     validationMap,
     disabledPaths,
+    enableDragAndDrop,
+    dnd: {
+      hooks: { useDrop },
+      ruleGroup: { isDragging, dragMonitorId, isOver, dropMonitorId, previewRef, dragRef, dropRef },
+    },
   } = schema;
   const { onGroupAdd, onGroupRemove, onPropChange, onRuleAdd, moveRule } = actions;
   const disabled = !!parentDisabled || !!disabledProp;
@@ -58,51 +63,6 @@ export const RuleGroup = ({
   }
 
   useDeprecatedProps('ruleGroup', !!ruleGroup);
-
-  const previewRef = useRef<HTMLDivElement>(null);
-  const dragRef = useRef<HTMLSpanElement>(null);
-  const dropRef = useRef<HTMLDivElement>(null);
-  const [{ isDragging, dragMonitorId }, drag, preview] = useDrag(
-    () => ({
-      type: DNDType.ruleGroup,
-      item: (): DraggedItem => ({ path }),
-      canDrag: !disabled,
-      collect: monitor => ({
-        isDragging: !disabled && monitor.isDragging(),
-        dragMonitorId: monitor.getHandlerId(),
-      }),
-    }),
-    [disabled, path]
-  );
-  const [{ isOver, dropMonitorId }, drop] = useDrop(
-    () => ({
-      accept: [DNDType.rule, DNDType.ruleGroup],
-      canDrop: (item: DraggedItem) => {
-        if (disabled) return false;
-        const parentItemPath = getParentPath(item.path);
-        const itemIndex = item.path[item.path.length - 1];
-        // Don't allow drop if 1) item is ancestor of drop target,
-        // 2) item is first child and is dropped on its own group header,
-        // or 3) the group is dropped on itself
-        return !(
-          isAncestor(item.path, path) ||
-          (pathsAreEqual(path, parentItemPath) && itemIndex === 0) ||
-          pathsAreEqual(path, item.path)
-        );
-      },
-      collect: monitor => ({
-        isOver: monitor.canDrop() && monitor.isOver(),
-        dropMonitorId: monitor.getHandlerId(),
-      }),
-      drop: (item: DraggedItem, _monitor) => !disabled && moveRule(item.path, [...path, 0]),
-    }),
-    [disabled, moveRule, path]
-  );
-  if (path.length > 0) {
-    drag(dragRef);
-    preview(previewRef);
-  }
-  drop(dropRef);
 
   const onCombinatorChange = (value: any) => {
     if (!disabled) {
@@ -181,6 +141,8 @@ export const RuleGroup = ({
     validationClassName,
     dndDragging
   );
+
+  const InlineCombinatorComponent = enableDragAndDrop ? InlineCombinatorDnD : InlineCombinator;
 
   return (
     <div
@@ -322,7 +284,7 @@ export const RuleGroup = ({
           return (
             <Fragment key={key}>
               {idx > 0 && !independentCombinators && showCombinatorsBetweenRules && (
-                <InlineCombinator
+                <InlineCombinatorComponent
                   options={combinators}
                   value={combinator}
                   title={translations.combinators.title}
@@ -337,10 +299,11 @@ export const RuleGroup = ({
                   path={thisPath}
                   disabled={thisPathDisabled}
                   independentCombinators={independentCombinators}
+                  useDrop={useDrop}
                 />
               )}
               {typeof r === 'string' ? (
-                <InlineCombinator
+                <InlineCombinatorComponent
                   options={combinators}
                   value={r}
                   title={translations.combinators.title}
@@ -355,6 +318,7 @@ export const RuleGroup = ({
                   path={thisPath}
                   disabled={thisPathDisabled}
                   independentCombinators={independentCombinators}
+                  useDrop={useDrop}
                 />
               ) : 'rules' in r ? (
                 <RuleGroupControlElement
