+++ packages/react-querybuilder/src/internal/hooks/index.ts
@@ -1,3 +1,7 @@
 export * from './useControlledOrUncontrolled';
 export * from './useDeprecatedProps';
+export * from './useInlineCombinatorDnD';
 export * from './usePrevious';
+export * from './useReactDnD';
+export * from './useRuleDnD';
+export * from './useRuleGroupDnD';
