+++ packages/react-querybuilder/src/internal/index.ts
@@ -1,6 +1,7 @@
 export * from './c';
 export * from './filterFieldsByComparator';
 export * from './generateID';
+export * from './getDndWrappers';
 export * from './getValidationClassNames';
 export * from './getValueSourcesUtil';
 export * from './messages';
