+++ packages/react-querybuilder/src/internal/types.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/consistent-type-imports */
 import type { API, FileInfo } from 'jscodeshift';
 
 export type LanguageParserCodeMod = (file: FileInfo, api: API) => string;
